[PaTH] Improving precision to fix NaNs

+++ tests/ops/test_path_attn.py
@@ -16,7 +16,7 @@ def naive_path_attn(q, k, v, w, beta, g, scale, BT=64):
     original_dtype = q.dtype
     HQ = q.shape[2]
     H = k.shape[2]
-    q, k, v, w, beta, g = map(lambda x: x.to(torch.float32).transpose(1, 2), [q, k, v, w, beta, g])
+    q, k, v, w, beta, g = map(lambda x: x.to(torch.float).transpose(1, 2), [q, k, v, w, beta, g])
     g_cumsum = g.cumsum(-1)
     q = q.unsqueeze(2).expand(-1, -1, HQ//HQ, -1, -1).flatten(1, 2)
     k = k.unsqueeze(2).expand(-1, -1, HQ//H, -1, -1).flatten(1, 2)
@@ -95,8 +95,8 @@ def test_parallel(
     q = torch.randn((B, T, HQ, D), dtype=dtype, device=device).requires_grad_(True)
     k = torch.randn((B, T, H, D), dtype=dtype, device=device).requires_grad_(True)
     v = torch.randn((B, T, H, D), dtype=dtype, device=device).requires_grad_(True)
-    w = F.normalize(torch.randn((B, T, H, D), dtype=dtype, device=device), dim=-1, p=2).requires_grad_(True)
-    beta = torch.empty((B, T, H), dtype=dtype, device=device).uniform_(1.5, 2.0).requires_grad_(True)
+    w = F.normalize(torch.randn((B, T, H, D), dtype=torch.float, device=device), dim=-1, p=2).requires_grad_(True)
+    beta = torch.empty((B, T, H), dtype=torch.float, device=device).uniform_(1.5, 2.0).requires_grad_(True)
     if use_forget_gate:
         g = torch.empty((B, T, HQ), dtype=torch.float, device=device).uniform_(
             0.95, 1).log().requires_grad_(True)
@@ -171,8 +171,8 @@ def test_parallel_varlen(
     q = torch.randn((1, T, HQ, D), dtype=dtype, device=device).requires_grad_(True)
     k = torch.randn((1, T, H, D), dtype=dtype, device=device).requires_grad_(True)
     v = torch.randn((1, T, H, D), dtype=dtype, device=device).requires_grad_(True)
-    w = F.normalize(torch.randn((1, T, H, D), dtype=dtype, device=device), dim=-1, p=2).requires_grad_(True)
-    beta = torch.rand((1, T, H), dtype=dtype, device=device).sigmoid().requires_grad_(True)
+    w = F.normalize(torch.randn((1, T, H, D), dtype=torch.float, device=device), dim=-1, p=2).requires_grad_(True)
+    beta = torch.rand((1, T, H), dtype=torch.float, device=device).sigmoid().requires_grad_(True)
     if use_forget_gate:
         g = torch.empty((1, T, HQ), dtype=torch.float, device=device).uniform_(0.95, 1).log().requires_grad_(True)
     else:
